requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
flatbuffers==24.12.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nltk==3.9.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
protobuf==5.29.2
Pygments==2.18.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
